HcOffOn
Put all three groups in the analysis

Adapted to new organization

+++ motor/motor_2ndlevelHcOffOn.m
@@ -16,84 +16,103 @@ spm('defaults', 'FMRI');
 %% Directories
 
 FDThresh = 0.3;
-ANALYSESDir = '/project/3022026.01/analyses/motor/fMRI_EventRelated_Main';
+ANALYSESDir = '/project/3022026.01/analyses/motor/DurAvg_ReAROMA_PMOD_TimeDer';
 % ANALYSESDir = '/project/3022026.01/analyses/motor/fMRI_EventRelated_BRCtrl';
-PITBIDSDir = '/project/3024006.01/bids';
-POMBIDSDir = '/project/3022026.01/bids';
-SubPIT = spm_BIDS(PITBIDSDir, 'subjects', 'task', 'motor');
-SubPOM = spm_BIDS(POMBIDSDir, 'subjects', 'task', 'motor');
-Sub = [SubPIT SubPOM];
+BIDSDir = '/project/3022026.01/pep/bids';
+BIDSDir_PIT = '/project/3022026.01/pep/bids_PIT';
+Sub = cellstr(spm_select('List', fullfile(ANALYSESDir), 'dir', '^sub-POM.*'));
 fprintf('Number of subjects processed: %i\n', numel(Sub))
 
 
 %% Selection
 
-SubSel = false(size(Sub));
-
-% Exclude participants without 1st-level analyses
+Sel = false(size(Sub));
+% Exclude participants without 1st-level analyses for visit 1
 for n = 1:numel(Sub)
-    if spm_select('List', fullfile(ANALYSESDir, ['sub-' Sub{n}], '1st_level'), '^con.*\.nii$')
-        fprintf('Including sub-%s with previous SPM output\n', Sub{n})
-        SubSel(n) = true;
-    else
-        fprintf('Excluding sub-%s without previous SPM output\n', Sub{n})
+    d = dir(fullfile(ANALYSESDir, Sub{n}, 'ses-Visit1*'));
+    for v = 1:length(d)
+        if spm_select('List', fullfile(d(v).folder, d(v).name, '1st_level'), '^con.*\.nii$')
+            Sel(n) = true;
+        else
+            fprintf('Excluding %s without previous SPM output for Visit1\n', Sub{n})
+        end
     end
 end
+Sub = Sub(Sel);
 
-% Take the last run for each participant if you're analzying POM data
-Run = num2str(ones(numel(Sub),1));
-for MultRunSub = spm_BIDS(POMBIDSDir, 'subjects', 'run','2', 'task','motor')
-    if ~isempty(MultRunSub)
-        fprintf('Altering run-number for sub-%s with run-2 data\n', char(MultRunSub))
-        index = strcmp(Sub,MultRunSub);
-        Run(index,1) = '2';
-    else
-        fprintf('No subjects with run-2 data\n')
+% Take the last run for each participant's session(s)
+Run = {};
+Session = {};
+Sub2 = {};
+for n = 1:numel(Sub)
+    Visit = cellstr(spm_select('List', fullfile(ANALYSESDir, Sub{n}), 'dir', 'ses-Visit1*'));
+    for v = 1:numel(Visit)
+        if strcmp(Visit{v}, 'ses-Visit1_PIT')
+            d = BIDSDir_PIT;
+        elseif strcmp(Visit{v}, 'ses-Visit1')
+            d = BIDSDir;
+        end
+        Sub2 = [Sub2; {Sub{n}}];
+        Session = [Session; {Visit{v}}];
+        Run = [Run; {num2str(FindLastRun(d, char(Sub{n}), 'ses-Visit1', 'motor', 'MB6'))}];
     end
 end
-for MultRunSub = spm_BIDS(POMBIDSDir, 'subjects', 'run','3', 'task','motor')
-    if ~isempty(MultRunSub)
-        fprintf('Altering run-number for sub-%s with run-3 data\n', char(MultRunSub))
-        index = strcmp(Sub,MultRunSub);
-        Run(index,1) = '3';
-    else
-        fprintf('No subjects with run-3 data\n')
-    end
-end
+SubInfo.Sub = Sub2;
+SubInfo.Session = Session;
+SubInfo.Run = Run;
 
-Sub = Sub(SubSel);
-Run = Run(SubSel);
+% Exclude Visit3
+Sel = true(size(SubInfo.Sub));
+for n = 1:numel(SubInfo.Sub)
+    if contains(SubInfo.Session{n}, 'ses-Visit3')
+        Sel(n) = false;
+        fprintf('Excluding Visit3 SPM output for %s \n', SubInfo.Sub{n})
+    end
+end
+fprintf('%i Visit3 excluded \n', sum(Sel == 0))
+SubInfo.Sub = SubInfo.Sub(Sel);
+SubInfo.Session = SubInfo.Session(Sel);
+SubInfo.Run = SubInfo.Run(Sel);
 
 %% Collect events.json and confound files
 
-JsonFiles = cell(size(Sub));
-ConfFiles = cell(size(Sub));
+JsonFiles = cell(size(SubInfo.Sub));
+ConfFiles = cell(size(SubInfo.Sub));
 
-for n = 1:numel(Sub)
-    if strncmp('PIT',Sub{n},3)
-        JsonFiles{n} = spm_select('FPList', fullfile(PITBIDSDir, ['sub-', Sub{n}], 'func'), ['.*task-motor_acq-MB6_run-' Run(n) '_events\.json$']);
+for n = 1:numel(SubInfo.Sub)
+    if contains(SubInfo.Session{n}, 'PIT')
+        JsonFiles{n} = spm_select('FPList', fullfile(BIDSDir_PIT, SubInfo.Sub{n}, 'ses-Visit1', 'beh'), ['.*task-motor_acq-MB6_run-' SubInfo.Run{n} '_events\.json$']);
     else
-        JsonFiles{n} = spm_select('FPList', fullfile(POMBIDSDir, ['sub-', Sub{n}], 'func'), ['.*task-motor_acq-MB6_run-' Run(n) '_events\.json$']);
+        JsonFiles{n} = spm_select('FPList', fullfile(BIDSDir, SubInfo.Sub{n}, 'ses-Visit1', 'beh'), ['.*task-motor_acq-MB6_run-' SubInfo.Run{n} '_events\.json$']);
     end
-    ConfFiles{n} = spm_select('FPList', fullfile(ANALYSESDir, ['sub-' Sub{n}]), ['^.*task-motor_acq-MB6_run-' Run(n) '.*desc-confounds_regressors2.mat$']);
+    ConfFiles{n} = spm_select('FPList', fullfile(ANALYSESDir, SubInfo.Sub{n}, SubInfo.Session{n}), ['^.*task-motor_acq-MB6_run-' SubInfo.Run{n} '_desc-confounds_regressors3.mat$']);
+    if isempty(ConfFiles{n})
+        ConfFiles{n} = spm_select('FPList', fullfile(ANALYSESDir, SubInfo.Sub{n}, SubInfo.Session{n}), ['^.*task-motor_acq-MB6_run-' SubInfo.Run{n} '_desc-confounds_regressors2.mat$']);
+    end
 end
 
 %% Group and handedness
 
-Group = strings(size(Sub));
-RespondingHand = strings(size(Sub));
+Group = strings(size(SubInfo.Sub));
+RespondingHand = strings(size(SubInfo.Sub));
 
-for n = 1:numel(Sub)
+for n = 1:numel(SubInfo.Sub)
     Json = fileread(JsonFiles{n});
     DecodedJson = jsondecode(Json);
     Group(n) = DecodedJson.Group.Value;
+    if contains(SubInfo.Session{n}, 'PIT')
+        d = BIDSDir_PIT;
+        Group(n) = FindGroup(d, SubInfo.Sub{n}, 'ses-Visit1', 'dwi');
+    else
+        Group(n) = 'PD_POM';
+    end
     RespondingHand(n) = DecodedJson.RespondingHand.Value;
 end
 
 %% Framewise displacement
 
-FD = zeros(size(Sub));
-for n = 1:numel(Sub)
+FD = zeros(size(SubInfo.Sub));
+for n = 1:numel(SubInfo.Sub)
     Confounds = spm_load(ConfFiles{n});
     FrameDisp = Confounds.R(:,strcmp(Confounds.names, 'framewise_displacement'));
     FrameDisp(isnan(FrameDisp)) = 0;
@@ -102,54 +121,54 @@ end
 
 %% Exclusion
 
-SubSel = false(size(Sub));
-
-for n = 1:numel(Sub)
+Sel = false(size(SubInfo.Sub));
+for n = 1:numel(SubInfo.Sub)
     if FD(n) < FDThresh
-        SubSel(n) = true;
+        Sel(n) = true;
     else
-        sprintf('Excluding %s (%s) due to %f mean framewise displacement (Threshold = %f)', Sub{n}, Group(n), FD(n), FDThresh)
+        sprintf('Excluding %s (%s) due to %f mean framewise displacement (Threshold = %f)', SubInfo.Sub{n}, Group(n), FD(n), FDThresh)
     end
 end
+FDexcluded = sum(Sel == 0);
+fprintf('%i participants excluded due to excessive motion \n', FDexcluded)
 
-Sub = Sub(SubSel);
-Group = Group(SubSel);
-RespondingHand = RespondingHand(SubSel);
-FD = FD(SubSel);
+SubInfo.Sub = SubInfo.Sub(Sel);
+SubInfo.Session = SubInfo.Session(Sel);
+SubInfo.Run = SubInfo.Run(Sel);
+Group = Group(Sel);
+RespondingHand = RespondingHand(Sel);
+FD = FD(Sel);
 
 %% Copy and flip
 
 ConList = {'con_0001' 'con_0002' 'con_0003' 'con_0004'};% 'con_0005'};
-InputFiles = cell(numel(Sub), numel(ConList));
+SessionList = {'ses-Visit1' 'ses-Visit1_PIT'};
 for c = 1:numel(ConList)
-    
     ConDir = fullfile(ANALYSESDir, 'Group', ConList{c});
     if ~exist(ConDir, 'dir')
         mkdir(ConDir);
     else
         delete(fullfile(ConDir, '*.*'));
     end
-    
-    for n = 1:numel(Sub)
-        InputConFile = fullfile(ANALYSESDir, ['sub-' Sub{n}], '1st_level', [ConList{c} '.nii']);
-        if strcmp(Group(n), 'PD_PIT')
-            OutputConFile = fullfile(ConDir, ['PD_PIT_' Sub{n} '_' ConList{c} '.nii']);
-        elseif strcmp(Group(n), 'PD_POM')
-            OutputConFile = fullfile(ConDir, ['PD_POM_' Sub{n} '_' ConList{c} '.nii']);
-        else
-            OutputConFile = fullfile(ConDir, ['HC_PIT_' Sub{n} '_' ConList{c} '.nii']);
-        end
-        copyfile(InputConFile, OutputConFile)
-        if strcmp(RespondingHand(n), 'Left') && Swap
-            fprintf('LR-swapping: %s\n', OutputConFile)
-			Hdr		  = spm_vol(OutputConFile);
-			Vol		  = spm_read_vols(Hdr);
-			Hdr.fname = spm_file(OutputConFile, 'suffix', 'LRswap');
-			spm_write_vol(Hdr, flipdim(Vol,1));		% LR is the first dimension in MNI space
-            delete(OutputConFile);
-            InputFiles{n,c} = Hdr.fname;
-        else
-            InputFiles{n,c} = OutputConFile;
+    for s = 1:numel(SessionList)
+        for n = 1:numel(SubInfo.Sub)
+            if strcmp(SessionList{s}, SubInfo.Session{n})
+                InputConFile = fullfile(ANALYSESDir, SubInfo.Sub{n}, SubInfo.Session{n}, '1st_level', [ConList{c} '.nii']);
+                if exist(InputConFile, 'file')
+                    OutputConFile = fullfile(ConDir, [char(Group(n)) '_' SubInfo.Sub{n} '_' SubInfo.Session{1} '_' ConList{c} '.nii']);
+                    copyfile(InputConFile, OutputConFile)
+                            if strcmp(RespondingHand(n), 'Left') && Swap
+                                fprintf('LR-swapping: %s\n', OutputConFile)
+                                Hdr		  = spm_vol(OutputConFile);
+                                Vol		  = spm_read_vols(Hdr);
+                                Hdr.fname = spm_file(OutputConFile, 'suffix', 'L2Rswap');
+                                spm_write_vol(Hdr, flipdim(Vol,1));		% LR is the first dimension in MNI space
+                                delete(OutputConFile);
+                            end
+                else
+                    fprintf('%s: Confile not available \n', SubInfo.Sub{n})
+                end
+            end
         end
     end
 end
@@ -162,34 +181,34 @@ Inputs{1,1} = {fullfile(ANALYSESDir, 'Group', 'HcOffOn x ExtInt2Int3Catch')};
 ExtHc = dir(fullfile(ANALYSESDir, 'Group', ConList{1}, 'HC_PIT*'));
 Inputs{2,1} = fullfile(ANALYSESDir, 'Group', ConList{1}, {ExtHc.name}');
 Int2Hc = dir(fullfile(ANALYSESDir, 'Group', ConList{2}, 'HC_PIT*'));
-Inputs{5,1} = fullfile(ANALYSESDir, 'Group', ConList{2}, {Int2Hc.name}');
+Inputs{3,1} = fullfile(ANALYSESDir, 'Group', ConList{2}, {Int2Hc.name}');
 Int3Hc = dir(fullfile(ANALYSESDir, 'Group', ConList{3}, 'HC_PIT*'));
-Inputs{8,1} = fullfile(ANALYSESDir, 'Group', ConList{3}, {Int3Hc.name}');
+Inputs{4,1} = fullfile(ANALYSESDir, 'Group', ConList{3}, {Int3Hc.name}');
 CatchHc = dir(fullfile(ANALYSESDir, 'Group', ConList{4}, 'HC_PIT*'));
-Inputs{11,1} = fullfile(ANALYSESDir, 'Group', ConList{4}, {CatchHc.name}');
+Inputs{5,1} = fullfile(ANALYSESDir, 'Group', ConList{4}, {CatchHc.name}');
 
 ExtPd = dir(fullfile(ANALYSESDir, 'Group', ConList{1}, 'PD_PIT*'));
-Inputs{3,1} = fullfile(ANALYSESDir, 'Group', ConList{1}, {ExtPd.name}');
+Inputs{6,1} = fullfile(ANALYSESDir, 'Group', ConList{1}, {ExtPd.name}');
 Int2Pd = dir(fullfile(ANALYSESDir, 'Group', ConList{2}, 'PD_PIT*'));
-Inputs{6,1} = fullfile(ANALYSESDir, 'Group', ConList{2}, {Int2Pd.name}');
+Inputs{7,1} = fullfile(ANALYSESDir, 'Group', ConList{2}, {Int2Pd.name}');
 Int3Pd = dir(fullfile(ANALYSESDir, 'Group', ConList{3}, 'PD_PIT*'));
-Inputs{9,1} = fullfile(ANALYSESDir, 'Group', ConList{3}, {Int3Pd.name}');
+Inputs{8,1} = fullfile(ANALYSESDir, 'Group', ConList{3}, {Int3Pd.name}');
 CatchPd = dir(fullfile(ANALYSESDir, 'Group', ConList{4}, 'PD_PIT*'));
-Inputs{12,1} = fullfile(ANALYSESDir, 'Group', ConList{4}, {CatchPd.name}');
+Inputs{9,1} = fullfile(ANALYSESDir, 'Group', ConList{4}, {CatchPd.name}');
 
 ExtPd = dir(fullfile(ANALYSESDir, 'Group', ConList{1}, 'PD_POM*'));
-Inputs{4,1} = fullfile(ANALYSESDir, 'Group', ConList{1}, {ExtPd.name}');
+Inputs{10,1} = fullfile(ANALYSESDir, 'Group', ConList{1}, {ExtPd.name}');
 Int2Pd = dir(fullfile(ANALYSESDir, 'Group', ConList{2}, 'PD_POM*'));
-Inputs{7,1} = fullfile(ANALYSESDir, 'Group', ConList{2}, {Int2Pd.name}');
+Inputs{11,1} = fullfile(ANALYSESDir, 'Group', ConList{2}, {Int2Pd.name}');
 Int3Pd = dir(fullfile(ANALYSESDir, 'Group', ConList{3}, 'PD_POM*'));
-Inputs{10,1} = fullfile(ANALYSESDir, 'Group', ConList{3}, {Int3Pd.name}');
+Inputs{12,1} = fullfile(ANALYSESDir, 'Group', ConList{3}, {Int3Pd.name}');
 CatchPd = dir(fullfile(ANALYSESDir, 'Group', ConList{4}, 'PD_POM*'));
 Inputs{13,1} = fullfile(ANALYSESDir, 'Group', ConList{4}, {CatchPd.name}');
 
-FD_HC_PIT = FD(strcmp(Group, 'HC_PIT'));
-FD_PD_PIT = FD(strcmp(Group, 'PD_PIT'));
-FD_PD_POM = FD(strcmp(Group, 'PD_POM'));
-Inputs{14,1} = [FD_HC_PIT FD_HC_PIT FD_HC_PIT FD_HC_PIT FD_PD_PIT FD_PD_PIT FD_PD_PIT FD_PD_PIT FD_PD_POM FD_PD_POM FD_PD_POM FD_PD_POM]';
+FD_HC_PIT = FD(strcmp(Group, 'HC_PIT'))';
+FD_PD_PIT = FD(strcmp(Group, 'PD_PIT'))';
+FD_PD_POM = FD(strcmp(Group, 'PD_POM'))';
+Inputs{14,1} = [FD_HC_PIT FD_HC_PIT  FD_HC_PIT FD_HC_PIT FD_PD_PIT FD_PD_PIT FD_PD_PIT FD_PD_PIT FD_PD_POM FD_PD_POM FD_PD_POM FD_PD_POM]';
 
 %% Run
